pyproject.toml
[project]
name = "what2watchnextbot"
version = "0.1.0"
description = "A Telegram bot to match you with a title you will watch tonight."
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "aiogram-loguru>=0.1.2",
    "aiogram[fast,i18n,redis]>=3.19.0",
    "alembic>=1.15.2",
    "loguru>=0.7.3",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy[postgresql-psycopgbinary]>=2.0.40",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.1",
    "logot[pytest]>=1.3.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-freezer>=0.4.9",
    "ruff>=0.11.5",
    "testfixtures>=8.3.0",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
required_plugins = [
    "pytest-asyncio",
    "logot",
]
pythonpath = "."
logot_capturer = "logot.loguru.LoguruCapturer"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # flake8-bugbear
    "B",
    # isort
    "I",
]

[tool.ruff.lint.isort]
known-third-party = ["alembic"]
